Replace handling of player deaths
As it turned out, the old death tracking logic was way too optimistic
about how the end leaderboard may look.
There's no way to guarantee that there's only a single winner as was
the original goal when implementing the game. During a single tick,
there can be multiple deaths and as such, more than one player can be
in the first place and this extends to any other place
in the leaderboard.

To minimize ties, we additionally take into account player's tail
length as we also already do when determining who wins a collision.